data_gathering/pyproject.toml
[tool.poetry]
authors = ["Andrii Sherman <watemerald@gmail.com>"]
description = ""
name = "fastapi-cloudrun-data-gathering"
version = "0.1.0"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.63.0"
uvicorn = "^0.13.4"
pydantic = "^1.8.1"
google-cloud-secret-manager = "^2.4.0"
google-cloud-storage = "^1.38.0"
httpx = "^0.18.1"

[tool.poetry.dev-dependencies]
autoflake = "*"
black = "^20.8b1"
isort = "^5.7.0"
mypy = "*"
pre-commit = "*"
pylint = "^2.6.0"
pytest = "*"
pytest-cov = "*"

[tool.isort]
profile = "black"

[tool.pytest.ini_options]

[tool.pylint.messages_control]
disable = "C0330, C0326"

[tool.pylint.format]
max-line-length = "88"

[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]
